# openapi.yaml
openapi: 3.1.0
info:
  title: Hookflow API
  description: |
    Enterprise-grade distributed webhook delivery system with at-least-once guarantees.

    ## Features
    - **Event Ingestion**: Send events via REST API with automatic fan-out to subscribed endpoints
    - **Reliable Delivery**: Automatic retries with exponential backoff (1m → 24h, 7 attempts)
    - **HMAC Signatures**: Secure webhook verification with SHA-256 signatures
    - **Multi-tenant**: Organization-based isolation with JWT authentication
    - **Rate Limiting**: Distributed rate limiting via Redis

    ## Authentication
    - **JWT Bearer Token**: For all operations (UI / user context)
    - **API Key**: For all project-scoped operations (`X-API-Key` header, SDK usage)
  contact:
    name: Hookflow
    url: https://github.com/vadymkykalo/webhook-platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: "{baseUrl}"
  description: Your Hookflow installation — the same origin the dashboard runs on
  variables:
    baseUrl:
      description: "Scheme and host of your installation, without a trailing slash"
      default: http://localhost:8080
security:
- bearerAuth: []
...
paths:
...
  /api/v1/auth/profile:
    put:
      tags:
      - Authentication
      summary: Update profile
      description: Updates the authenticated user's profile information
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
        required: true
      responses:
        "200":
          description: Profile updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
      security:
      - bearerAuth: []
...
  /api/v1/auth/verify-email:
    post:
      tags:
      - Authentication
      summary: Verify email
      description: Verifies user email with the token sent to their email
      operationId: verifyEmail
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Email verified successfully
        "400":
          description: Invalid or expired token
  /api/v1/auth/switch-organization:
    post:
      tags:
      - Authentication
      summary: Switch organization
      description: "Re-issues an access token scoped to another organization the caller\
        \ belongs to. The refresh token is untouched and no other session is affected;\
        \ the new token's role comes from the membership in the target organization,\
        \ not from the token being replaced."
      operationId: switchOrganization
      parameters:
      - name: refresh_token
        in: cookie
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchOrganizationRequest"
        required: true
      responses:
        "200":
          description: Access token re-issued for the target organization
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: "Not authenticated, or no live session"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "403":
          description: Caller is not a member of that organization
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
      security:
      - bearerAuth: []
  /api/v1/auth/sessions/revoke-all:
    post:
      tags:
      - Authentication
      summary: Sign out everywhere
      description: "Ends every session for the authenticated user, including the one\
        \ making the call."
      operationId: revokeAllSessions
      responses:
        "204":
          description: All sessions revoked
        "401":
          description: Not authenticated
      security:
      - bearerAuth: []
  /api/v1/auth/reset-password:
    post:
      tags:
      - Authentication
      summary: Reset password
      description: Resets user password using the token from the reset email
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "200":
          description: Password reset successfully
        "400":
          description: Invalid or expired token
  /api/v1/auth/resend-verification:
    post:
      tags:
      - Authentication
      summary: Resend verification email
      description: Sends a new verification email to the user
      operationId: resendVerification
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Verification email sent
        "400":
          description: Email already verified
        "404":
          description: User not found
        "429":
          description: Too many requests
  /api/v1/auth/register:
    post:
      tags:
      - Authentication
      summary: Register new user
      description: Creates a new user account and organization
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: User registered successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request or email already exists
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh token
      description: Exchanges a valid refresh token for new access and refresh tokens
      operationId: refreshToken
      parameters:
      - name: refresh_token
        in: cookie
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Tokens refreshed successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/auth/logout:
    post:
      tags:
      - Authentication
      summary: Logout
      description: Revokes access and refresh tokens
      operationId: logout
      parameters:
      - name: refresh_token
        in: cookie
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: Logged out successfully
        "401":
          description: Not authenticated
      security:
      - bearerAuth: []
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      description: Authenticates user and returns JWT token
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Login successful
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/auth/forgot-password:
    post:
      tags:
      - Authentication
      summary: Forgot password
      description: Sends a password reset email to the user
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "200":
          description: "If the email exists, a reset link has been sent"
        "429":
          description: Too many requests
...
  /api/v1/auth/change-password:
    post:
      tags:
      - Authentication
      summary: Change password
      description: Changes the authenticated user's password
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "200":
          description: Password changed successfully
        "400":
          description: Current password incorrect or validation failed
        "401":
          description: Not authenticated
      security:
      - bearerAuth: []
...
  /api/v1/auth/sessions:
    get:
      tags:
      - Authentication
      summary: List active sessions
      description: Returns every live sign-in for the authenticated user — browser
        sessions and CLI device-code grants alike — with the session making the request
        flagged as current. No token material is returned.
      operationId: listSessions
      parameters:
      - name: refresh_token
        in: cookie
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Sessions returned
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionResponse"
        "401":
          description: Not authenticated
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionResponse"
      security:
      - bearerAuth: []
  /api/v1/auth/me:
    get:
      tags:
      - Authentication
      summary: Get current user
      description: Returns information about the authenticated user
      operationId: getCurrentUser
      responses:
        "200":
          description: User info returned
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
        "401":
          description: Not authenticated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
      security:
      - bearerAuth: []
...
  /api/v1/auth/sessions/{sessionId}:
    delete:
      tags:
      - Authentication
      summary: Revoke a session
      description: "Signs one device out. The session's refresh token stops working\
        \ and so does the access token it already issued, rather than surviving until\
        \ its own expiry."
      operationId: revokeSession
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Session revoked
        "401":
          description: Not authenticated
        "404":
          description: No such session for this user
      security:
      - bearerAuth: []
components:
  schemas:
...
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
          maxLength: 255
          minLength: 0
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        fullName:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - PENDING_VERIFICATION
          - DISABLED
...
    SwitchOrganizationRequest:
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
      required:
      - organizationId
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        emailVerified:
          type: boolean
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
        newPassword:
          type: string
          maxLength: 128
          minLength: 8
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&\\-_#^()+=])[A-Za-z\\\
            d@$!%*?&\\-_#^()+=]{8,}$"
      required:
      - newPassword
      - token
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          maxLength: 128
          minLength: 8
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&\\-_#^()+=])[A-Za-z\\\
            d@$!%*?&\\-_#^()+=]{8,}$"
        fullName:
          type: string
          maxLength: 255
          minLength: 0
        organizationName:
          type: string
          maxLength: 100
          minLength: 2
        captchaToken:
          type: string
          maxLength: 4096
          minLength: 0
      required:
      - email
      - organizationName
      - password
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
      - refreshToken
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT token obtained from /api/v1/auth/login
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      description: Project API key for event ingestion
      name: X-API-Key
      in: header
    platformAdminToken:
      type: apiKey
      description: "Cluster-operator credential (PLATFORM_ADMIN_TOKEN env var), independent\
        \ of tenant org membership — required for cross-tenant admin endpoints"
      name: X-Platform-Admin-Token
      in: header